Move customers.ts, profiles.ts, staff-invitations.ts, stores.ts to schema/app
