+++ backend/drizzle.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from "drizzle-kit";
 
 export default defineConfig({
     dialect: "postgresql",
-    schema: ["./src/db/schema/stores.ts", "./src/db/schema/customers.ts", "./src/db/schema/profiles.ts", './src/db/schema/staff-invitations.ts'],
+    schema: "./src/db/schema/app",
     out: "./sql/01_app",
     dbCredentials: {
         host: "localhost",
