+++ backend/src/db/schema/app/stores.ts
@@ -7,7 +7,7 @@ import {
   timestamp,
   varchar,
 } from "drizzle-orm/pg-core";
-import { CUID_LENGTH } from "../constants";
+import { CUID_LENGTH } from "../../constants";
 import { createId } from "@paralleldrive/cuid2";
 import { relations } from "drizzle-orm";
